# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Orbyn API
  version: "1"
  summary: Tasks, events, time blocks and webhooks, for scripts and automation tools.
  description: |
    The part of Orbyn's API that automation tools (Zapier, Make, n8n, your own
    scripts) use. Nothing here sends your data anywhere until you connect it
    (a key, a webhook or a feed link), and each can be turned off at any time.
    The full reference, including planning, teams and booking pages, is
    docs/api.md.

    **Authentication.** Create a personal API key in Settings → Connections
    (or `POST /me/api-keys`) and send it as `Authorization: Bearer ok_…`. A key
    acts as you for items, pages and the calendar. It is refused (`403`) on
    the admin console and wherever it could take over or change the account:
    creating keys or removing any key but itself, webhooks, chat delivery,
    sessions, two-step verification, passkeys, export, deleting the account,
    account settings (`PUT /me`, the public profile, privacy, time zone,
    consent, the email-to-task address and the calendar feed link), push
    devices, agent keys, and the assistant. Session tokens from
    `POST /auth/login` work everywhere except `/mcp`.

    **Agents.** AI agents connect to the MCP address (docs/mcp.md) with an
    agent key (`oak_…`, made with `POST /me/agent-keys`). Agent credentials
    are refused (`401`) by every route described here.

    **Errors** are `{ "message": "..." }` with the status codes below. `404`
...

...

...

...
servers:
...
security:
  - bearer: []
...
paths:
...
  /me/agents:
    get:
      tags: [Auth]
      summary: Your connected agents and the MCP address (not with an API key)
      responses:
        "200":
          description: "{ mcp_url, legacy_keys_until, grants }"
          content:
            application/json:
              schema: { type: object }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /me/agent-keys:
    post:
      tags: [Auth]
      summary: Make an agent key for an AI agent (10 a minute; not with an API key)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, maxLength: 80 }
                access: { type: string, enum: [read, suggest, write] }
                personal: { type: boolean }
                team_ids: { type: array, items: { type: string, format: uuid } }
                expires_in_days: { type: integer, minimum: 1, maximum: 365 }
                hide_outside_content: { type: boolean }
      responses:
        "201":
          description: "{ grant, key }: the key (oak_…) is shown once"
          content:
            application/json:
              schema: { type: object }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /me/agents/{id}:
    delete:
      tags: [Auth]
      summary: Revoke a connected agent
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses:
        "204": { description: Revoked }
        "404": { $ref: "#/components/responses/NotFound" }
  /me/agents/{id}/restore:
    post:
      tags: [Auth]
      summary: Restore a connected agent Orbyn paused for misbehaving
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses:
        "204": { description: Restored }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /me/agents/{id}/activity:
    get:
      tags: [Auth]
      summary: What one connected agent did
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses:
        "200":
          description: Activity, newest first
          content:
            application/json:
              schema: { type: array, items: { type: object } }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: A personal API key (`ok_…`) or a session token.
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    Unauthorized:
      description: Missing, expired or revoked token
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Not allowed, for example an API key on a route that manages the account
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Not found, including records that aren't yours
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    RateLimited:
      description: Too many requests
      headers:
        Retry-After:
          description: Seconds to wait before trying again
          schema: { type: integer }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [message]
      properties:
        message: { type: string }
...